Changed project structrure
Changed project folder structure so that I can have eslint for the express part of the project (I can't install eslint in a parent folder of an app generated by create-react-app or it throws an error).
