+++ api/.eslintrc.js
@@ -1,0 +1,15 @@
+module.exports = {
+  env: {
+    es2020: true,
+  },
+  extends: [
+    'airbnb-base',
+  ],
+  parserOptions: {
+    ecmaVersion: 11,
+    sourceType: 'module',
+  },
+  rules: {
+    "linebreak-style": 0,
+  },
+};
