+++ package.json
@@ -1,32 +1,25 @@
 {
-  "name": "react-tutorial",
+  "name": "github",
   "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
+  "description": "App that resizes pics as part of a tutorial.",
+  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js",
-    "server": "nodemon server.js",
+    "api": "npm run server --prefix api",
     "client": "npm start --prefix client",
-    "dev": "concurrently \"npm run server\" \"npm run client\""
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "concurrently": "^5.2.0",
-    "express": "^4.17.1"
+    "dev": "concurrently \"npm run api\" \"npm run client\""
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/SofiaKef/tutorial-app.git"
   },
+  "author": "",
+  "license": "ISC",
   "bugs": {
     "url": "https://github.com/SofiaKef/tutorial-app/issues"
   },
   "homepage": "https://github.com/SofiaKef/tutorial-app#readme",
-  "devDependencies": {
-    "eslint-config-airbnb-base": "^14.2.0",
-    "eslint-plugin-import": "^2.22.0",
-    "nodemon": "^2.0.4"
+  "dependencies": {
+    "concurrently": "^5.2.0"
   }
 }
